Us6 and 7 complete

+++ db/migrate/20230302222223_create_bulk_discounts.rb
@@ -2,7 +2,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[5.2]
   def change
     create_table :bulk_discounts do |t|
       t.integer :quantity_threshold
-      t.integer :percentage_discount
+      t.float :percentage_discount
     end
   end
 end
